updating compile.sh for cleaning

+++ bin/compile.sh
@@ -28,12 +28,16 @@ foreach type ($types)
 
    set d = `date +"%Y-%m-%d %H:%M:%S"`
 
-   printf "$d %-14s %-14s" "${platform}" "cleaning..."
-#     scons arch=$arch type=$type -c >& /dev/null
-   rm -f $dir/*unit >& /dev/null
-   rm -f bin/$type/* >& /dev/null
-   printf "done\n"
+   if ($type == "clean") then
 
+     printf "$d %-14s %-14s" "${platform}" "cleaning..."
+     scons arch=$arch type=serial -c >& /dev/null
+     scons arch=$arch type=mpi    -c >& /dev/null
+     scons arch=$arch type=charm  -c >& /dev/null
+     rm -f $dir/*unit >& /dev/null
+     rm -f bin/$type/* >& /dev/null
+     printf "done\n"
+   else
 
    # COMPILE
 
@@ -82,6 +86,7 @@ foreach type ($types)
       endif
    endif
 
+   endif
    printf "\n"
